Take vertices in as floats (unquantized)
* No reverse of quantization on spicy draco decode

+++ src/draco/psy/psy_draco_encoder.h
@@ -51,7 +51,7 @@ public:
 
     bool IsTexCoordInfoCompressing() const;
 
-    eStatus Run(const int16_t* pQuantizedVertices,
+    eStatus Run(const float* pQuantizedVertices,
                 const size_t vertexStride,
                 const size_t verticesCount,
                 const float decodeMultiplier,
